+++ fluent_pages/adminui/urlnodechildadmin.py
@@ -173,7 +173,7 @@ class UrlNodeChildAdmin(MultiSiteAdminMixin, PolymorphicMPTTChildModelAdmin, Tra
         fields = super(UrlNodeChildAdmin, self).get_readonly_fields(request, obj)
         if obj is not None:
             # Edit screen
-            if obj.get_available_languages().count() >= 2 \
+            if len(obj.get_available_languages()) >= 2 \
             and not self.has_change_shared_fields_permission(request, obj):
                 # This page is translated in multiple languages,
                 # language team is only allowed to update their own language.