+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes in git
 
 * Added ``draft`` CSS class to unpublished menu items that are only visible for staff members.
 * Added ``FluentPagesConfig`` to use Django 1.7 appconfigs.
+* Added multiple fallback language support for django-parler_ 1.5.
 * Improve performance of ``PageSitemap`` for sites with a lot of pages.
 * Fixed Django 1.8 issues in the "Change Page" view.
 * Fixed silent behavior of using ``.parent_site()`` too late in an already filtered queryset.